requirements.txt
# LAN Communication Application Dependencies

# GUI Framework
PySide6>=6.5.0

# Media capture (optional for production - stubs provided)
opencv-python>=4.8.0
numpy>=1.24.0

# Audio capture
sounddevice>=0.4.6

# Packaging
pyinstaller>=5.13.0

# Screen capture
mss>=9.0.0
Pillow>=10.0.0

# Optional utilities
# python-magic>=0.4.27
